agent 汇报cpu 内存使用情况

+++ src/flags.cc
@@ -17,6 +17,6 @@ std::string FLAGS_container = "cmd";// cmd or cgroup
 
 int FLAGS_agent_curl_recv_buffer_size = 1024 * 10;
 
-int FLAGS_cpu_num = 4;
+double FLAGS_cpu_num = 4;
 int FLAGS_mem_gbytes = 32;
 /* vim: set expandtab ts=4 sw=4 sts=4 tw=100: */
